Dep updates and profile refactors
